refactor: separate some grid functionality into custom hooks
